Refactor to extract service classes
