+++ app/services/dqt/find_dead_jobs.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module Dqt
+  class FindDeadJobs < BaseFindJobs
+    def sidekiq_class
+      Sidekiq::DeadSet
+    end
+  end
+end
