+++ app/services/dqt/find_retry_jobs.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module Dqt
+  class FindRetryJobs < BaseFindJobs
+    def sidekiq_class
+      Sidekiq::RetrySet
+    end
+  end
+end
